Fixing issues with player class
