+++ spec/player_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../lib/player_2.rb'
+require_relative '../lib/player.rb'
 
 RSpec.describe 'Player' do
   describe '#initialize' do
